+++ lib/assets/javascripts/new-dashboard/components/DatasetsList.vue
@@ -123,7 +123,8 @@ export default {
   data () {
     return {
       isScrollPastHeader: false,
-      selectedDatasets: []
+      selectedDatasets: [],
+      lastCheckedItem: null
     };
   },
   created: function () {
@@ -187,12 +188,13 @@ export default {
       this.$emit('applyOrder', orderParams);
     },
     toggleSelected ({ dataset, isSelected, event }) {
-      if (event.shiftKey) {
+      if (this.selectedDatasets.length && event.shiftKey) {
         this.doShiftClick(dataset);
         return;
       }
 
       if (isSelected) {
+        this.lastCheckedItem = dataset;
         this.selectedDatasets.push(dataset);
         return;
       }
@@ -201,7 +203,7 @@ export default {
     },
     doShiftClick (dataset) {
       const datasetsArray = [...Object.values(this.datasets)];
-      this.selectedDatasets = shiftClick(datasetsArray, this.selectedDatasets, dataset);
+      this.selectedDatasets = shiftClick(datasetsArray, this.selectedDatasets, dataset, this.lastCheckedItem || dataset);
     },
     selectAll () {
       this.selectedDatasets = [...Object.values(this.$store.state.datasets.list)];